backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
google-generativeai==0.3.2
python-dotenv==1.0.1
pydantic==2.6.4
chromadb==0.4.22